Quaternions (no rotation yet)

+++ 3D/src/Orientable.java
@@ -1,9 +1,9 @@
 import java.awt.Graphics;
 
 public interface Orientable{
-    public int x();
-    public int y();
-    public int z();
+    public double x();
+    public double y();
+    public double z();
     public int obscures(Orientable o);
     public int getDist();
     public void draw2D(Graphics g);
